ref: Move symbolication service tests
Moves some in-module tests around so they are now integration tests.
