+++ crates/symbolicator-service/tests/integration/main.rs
@@ -1,0 +1,6 @@
+// See <https://matklad.github.io/2021/02/27/delete-cargo-integration-tests.html>
+
+pub mod process_minidump;
+pub mod symbolication;
+#[macro_use]
+pub mod utils;
