+++ crates/symbolicator-service/tests/integration/snapshots/CAUTION.md
