+++ crates/symbolicator-test/src/lib.rs
@@ -42,7 +42,7 @@ pub use tempfile::TempDir;
 ///    other logs (such as actix or symbolic).
 pub fn setup() {
     fmt()
-        .with_env_filter(EnvFilter::new("symbolicator=trace"))
+        .with_env_filter(EnvFilter::new("symbolicator-service=trace"))
         .with_target(false)
         .pretty()
         .with_test_writer()
